# lazyCallReq: Only parse transport headers once