--- a/relay_messages_benchmark_test.go
+++ b/relay_messages_benchmark_test.go
@@ -1,0 +1,49 @@
+// Copyright (c) 2015 Uber Technologies, Inc.
+
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in
+// all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+// THE SOFTWARE.
+
+package tchannel
+
+import (
+	"fmt"
+	"io/ioutil"
+	"testing"
+)
+
+func BenchmarkCallReqFrame(b *testing.B) {
+	cr := reqHasAll.req()
+	f := cr.Frame
+
+	var service, caller, method string
+
+	b.ResetTimer()
+	for i := 0; i < b.N; i++ {
+		cr := newLazyCallReq(f)
+
+		// Multiple calls due to peer selection, stats, etc.
+		for i := 0; i < 3; i++ {
+			service = cr.Service()
+			caller = cr.Caller()
+			method = cr.Method()
+		}
+	}
+	b.StopTimer()
+
+	fmt.Fprint(ioutil.Discard, service, caller, method)
+}