I'm parsing the frame bytes explicitly instead of using `typed` to avoid extra allocations from creating strings for each header.

Benchmark comparison:
```
benchmark                   old ns/op     new ns/op     delta
BenchmarkCallReqFrame-8     1046          214           -79.54%

benchmark                   old allocs     new allocs     delta
BenchmarkCallReqFrame-8     24             5              -79.17%

benchmark                   old bytes     new bytes     delta
BenchmarkCallReqFrame-8     120           48            -60.00%
```