@prashantv 2016-06-13
Updated branch + benchmarks in original PR description, PTAL